% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifEarth.R
\docType{data}
\name{LearnerClassifEarth}
\alias{LearnerClassifEarth}
\alias{mlr_learners_classif.earth}
\title{MARS (Multivariate Adaptive Regression Splines) Classification
(binary)}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerClassif}.}
\usage{
LearnerClassifEarth
}
\description{
A \link{LearnerClassif} for MARS implemented in earth::earth()] in package
\CRANpkg{earth}.
This package is derived from mda::mars()] in package \CRANpkg{mda}.
The classification framework is derived via the the arguments of earth.
In particular we set the \code{glm} argument to
\code{list(family = "binomial")}
Methods for variance estimations have not been implemented so far.
}
\section{Construction}{
\preformatted{LearnerClassifRanger$new()
mlr3::mlr_learners$get("classif.earth")
mlr3::lrn("classif.earth")
}
}

\references{
Stephen Milborrow (2014)
Earth: multivariate adaptive regression spline models
R package version 3
\url{https://cran.r-project.org/web/packages/earth/earth.pdf}

Jerome H. Friedman (1991)
Multivariate Adaptive Regression Splines
The Annals of Statistics
\url{https://projecteuclid.org/download/pdf_1/euclid.aos/1176347963}
}
\keyword{datasets}
