% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerRegrEarth.R
\docType{data}
\name{LearnerRegrEarth}
\alias{LearnerRegrEarth}
\alias{mlr_learners_regr.earth}
\title{MARS (Multivariate Adaptive Regression Splines)}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerRegr}.}
\usage{
LearnerRegrEarth
}
\description{
A \link{LearnerRegr} for MARS implemented in earth::earth()] in package
\CRANpkg{earth}.
This package is derived from mda::mars()] in package \CRANpkg{mda}.
Methods for variance estimations have not been implemented so far.
}
\section{Construction}{
\preformatted{LearnerRegrEarth$new()
mlr3::mlr_learners$get("regr.earth")
mlr3::lrn("regr.earth")
}
}

\references{
Stephen Milborrow (2014)
Earth: multivariate adaptive regression spline models
R package version 3
\url{https://cran.r-project.org/web/packages/earth/earth.pdf}

Jerome H. Friedman (1991)
Multivariate Adaptive Regression Splines
The Annals of Statistics
\url{https://projecteuclid.org/download/pdf_1/euclid.aos/1176347963}
}
\keyword{datasets}
