% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.earth-package}
\alias{mlr3learners.earth}
\alias{mlr3learners.earth-package}
\title{mlr3learners.earth: Multivariate Adaptive Regression Splines for mlr3 (earth package)}
\description{
Wrapper for a classification and regression learner from the earth 
  package (CRAN) to be used in mlr3.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr3learners}
  \item \url{https://github.com/mlr3learners/mlr3learners.earth}
}

}
\author{
\strong{Maintainer}: Philipp Kopper \email{kopperphilipp@gmail.com}

}
